requirements.txt
requests==2.31.0
pytest==7.4.3
allure-pytest==2.13.2
PyYAML==6.0.1
python-dotenv==1.0.0 